requirements.txt
# Core
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Google AI
google-generativeai>=0.8.0

# Image Processing
Pillow>=10.0.0

# Lineage & Royalties
networkx>=3.2.0

# Database
sqlalchemy>=2.0.0
asyncpg>=0.29.0

# Storage
google-cloud-storage>=2.13.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
